Restrict redacted-line construction and retain explicit continuations
