+++ Packages/GuesthouseCore/Sources/GuesthouseCore/Redaction/Redactor+API.swift
@@ -129,7 +129,7 @@ extension Redactor {
         // completes, and only a scan arms that, so the rules run here for their state alone.
         if authorizationContinuation {
             state.quotedValue = Self.unterminatedQuote(in: text[...], kind: "authorization")
-            if authorizationWasAwaitingValue, Self.valueStartsOnNextLine(text[...]) {
+            if Self.valueStartsOnNextLine(text[...]) {
                 state.authorizationValueIsOnTheNextLine = true
             }
             Self.armPendingContexts(from: text, state: &state)
@@ -204,5 +204,39 @@ extension Redactor {
         }
         appendRedacted(text[lineStart...])
         return result
+    }
+}
+
+/// A line of text that has passed through `Redactor`.
+///
+/// Log sinks, the XPC event stream, and diagnostics exports accept only this type, so an
+/// unredacted `String` cannot reach them by accident. The only ways to obtain one are
+/// `Redactor` and a compile-time string literal.
+struct RedactedLine: Hashable, Sendable, CustomStringConvertible {
+    let text: String
+
+    fileprivate init(_ text: String) {
+        self.text = text
+    }
+
+    /// For fixed messages that contain nothing to redact, such as `"Started"`.
+    init(literal: StaticString) {
+        text = literal.description
+    }
+
+    var description: String { text }
+}
+
+extension RedactedLine: Codable {
+    /// Decoded text is redacted again: a serialized or XPC-provided value is not trusted to
+    /// have passed through `Redactor` on the other side. The line that gave a device code its
+    /// context did not survive serialization, so codes are redacted unconditionally here.
+    init(from decoder: any Decoder) throws {
+        text = Redactor().redact(untrusted: try decoder.singleValueContainer().decode(String.self))
+    }
+
+    func encode(to encoder: any Encoder) throws {
+        var container = encoder.singleValueContainer()
+        try container.encode(text)
     }
 }
